ran a clean up, and did some cleaning of my own, changed other stuff too

+++ src/org/usfirst/frc/team649/robot/subsystems/CameraSubsystem.java
@@ -13,16 +13,6 @@ import java.util.ArrayList;
 import java.util.List;
 
 
-import org.opencv.core.Core;
-import org.opencv.core.CvType;
-import org.opencv.core.Mat;
-import org.opencv.core.MatOfPoint;
-import org.opencv.core.Rect;
-import org.opencv.core.Scalar;
-import org.opencv.core.Size;
-import org.opencv.highgui.VideoCapture;
-import org.opencv.imgproc.Imgproc;
-import org.opencv.imgproc.Moments;
 import org.usfirst.frc.team649.robot.RobotMap;
 import org.usfirst.frc.team649.robot.RobotMap.Drivetrain;
 import org.usfirst.frc.team649.robot.util.Center;
@@ -178,7 +168,7 @@ public class CameraSubsystem extends Subsystem {
 	        	center = new Center(mu.get_m10()/mu.get_m00(), mu.get_m01()/mu.get_m00());
 	        	
 	        	String[] keys = new String[]{"Obj Center X: ", "Obj Center Y: ", "Obj Area: ", "Obj width: ", "Obj height: ", "Obj Distance: "}; 
-	        	Object[] elements = new Object[]{(Double)center.x, (Double)center.y, (Double)Imgproc.contourArea(contours.get(largest)), (Integer)r.width, (Integer)r.height, (Double)dist};
+	        	Object[] elements = new Object[]{center.x, center.y, (Double)Imgproc.contourArea(contours.get(largest)), (Integer)r.width, (Integer)r.height, dist};
 //	        	"Obj 0 Center X: ", mu.get_m10()/mu.get_m00());
 //	        	"Obj 0 Center Y: ", mu.get_m01()/mu.get_m00());
 //	        	"Obj 0 Area: ", Imgproc.contourArea(contours.get(largest)));
